@boosty 2015-09-09
Even faster when using the ! versions of `tr` and `map` :wink: 

```ruby
require 'benchmark/ips'

KEY = :'HALLO_WELT-Wie_gEHTs'

def convert_key(key)
  key.to_s.split(/_|-/).map { |segment| segment.capitalize }.join("-")
end

def convert_key_freeze(key)
  key.to_s.split(/_|-/).map { |segment| segment.capitalize.freeze }.join("-".freeze)
end

def convert_key_tr_freeze(key)
  key.to_s.tr('-'.freeze,'_'.freeze).split('_'.freeze).map { |segment| segment.capitalize.freeze }.join("-".freeze)
end

def convert_key_tr_freeze_bang(key)
  key.to_s.tr('-'.freeze,'_'.freeze).split('_'.freeze).map! { |segment| segment.capitalize.freeze }.join("-".freeze)
end

def convert_key_tr_freeze_bang2(key)
  k = key.to_s
  k.tr!('-'.freeze,'_'.freeze)
  k.split('_'.freeze).map! { |segment| segment.capitalize.freeze }.join("-".freeze)
end

def convert_key_tr_no_freeze(key)
  key.to_s.tr('-','_').split('_').map { |segment| segment.capitalize }.join("-")
end


Benchmark.ips do |x|
  x.report('no freeze') { convert_key(KEY) }
  x.report('freeze') { convert_key_freeze(KEY) }
  x.report('tr_freeze') { convert_key_tr_freeze(KEY) }
  x.report('tr_freeze_bang') { convert_key_tr_freeze_bang(KEY) }
  x.report('tr_freeze_bang2') { convert_key_tr_freeze_bang2(KEY) }
  x.report('tr_no_freeze') { convert_key_tr_no_freeze(KEY) }
  x.compare!
end
```

```
Calculating -------------------------------------
           no freeze    18.757k i/100ms
              freeze    18.680k i/100ms
           tr_freeze    24.012k i/100ms
      tr_freeze_bang    25.883k i/100ms
     tr_freeze_bang2    26.342k i/100ms
        tr_no_freeze    23.298k i/100ms
-------------------------------------------------
           no freeze    219.361k (± 3.3%) i/s -      1.107M
              freeze    217.668k (± 3.2%) i/s -      1.102M
           tr_freeze    296.556k (± 3.0%) i/s -      1.489M
      tr_freeze_bang    315.970k (± 3.7%) i/s -      1.579M
     tr_freeze_bang2    319.472k (± 5.0%) i/s -      1.607M
        tr_no_freeze    276.707k (± 3.2%) i/s -      1.398M

Comparison:
     tr_freeze_bang2:   319472.3 i/s
      tr_freeze_bang:   315970.3 i/s - 1.01x slower
           tr_freeze:   296556.0 i/s - 1.08x slower
        tr_no_freeze:   276706.5 i/s - 1.15x slower
           no freeze:   219361.2 i/s - 1.46x slower
              freeze:   217668.4 i/s - 1.47x slower
```