@Aqualon 2015-09-09
@boosty but not working with `tr!` because the values given to the convert_key methods are string keys and those are automatically frozen by ruby. But the `map!` is faster. Will adapt this.