@Aqualon 2015-09-09
Squashed all the changes into one commit.

On sandbox memory consumption for /lib/typhoeus/normalized_header_hash.rb went from 14392  allocated objects down to 11134 (for displaying one story in feedy). Allocations were measured with memory_profiler gem.