@boosty 2015-09-10
Also "thanks" from my side! I will release a new gem version soon. But first I want to throw out the old-fashioned jeweler-based gem setup :wink: